fix: 1등 당첨이 여러번 되었을때의 오버플로우 문제 해결

+++ src/main/java/lotto/controller/LottoController.java
@@ -35,7 +35,7 @@ public class LottoController {
 
     private Money getMoney() {
         InputHandler inputHandler = new InputHandler(new Validator());
-        int givenMoney = inputHandler.getInputForPurchaseMoney();
+        long givenMoney = inputHandler.getInputForPurchaseMoney();
         return new Money(givenMoney);
     }
