+++ src/main/java/lotto/domain/Money.java
@@ -1,40 +1,40 @@
 package lotto.domain;
 
 public class Money {
-    private static final int ZER0_MONEY = 0;
-    private static final int LOTTO_PRICE = 1000;
-    private int money;
-    private int spentMoney;
-    private int earnedMoney;
+    private static final long ZER0_MONEY = 0;
+    private static final long LOTTO_PRICE = 1000;
+    private long money;
+    private long spentMoney;
+    private long earnedMoney;
 
     public Money() {
         this(ZER0_MONEY);
     }
 
-    public Money(int money) {
+    public Money(long money) {
         this.money = money;
     }
 
-    public int getTicket() {
-        int ticketCount = money / LOTTO_PRICE;
+    public long getTicket() {
+        long ticketCount = money / LOTTO_PRICE;
         useMoney(money);
         return ticketCount;
     }
 
-    public int getEarnedMoney() {
+    public long getEarnedMoney() {
         return earnedMoney;
     }
 
-    public int getSpentMoney() {
+    public long getSpentMoney() {
         return spentMoney;
     }
 
-    public void addMoney(int amount) {
+    public void addMoney(long amount) {
         money += amount;
         earnedMoney += amount;
     }
 
-    private void useMoney(int amount) {
+    private void useMoney(long amount) {
         money -= amount;
         spentMoney += amount;
     }
